Use enum for light occluder shapes
This will allow us to sanely expand our API as we introduce more
occluder shapes.
